Projet terminé

+++ src/app/list-plat/list-plat.component.ts
@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Plat } from '../models/plat';
 
 import { PlatService } from '../plat.service';
-
+import { RoleGuardService } from '../role-guard.service';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-list-plat',
   templateUrl: './list-plat.component.html',
@@ -10,7 +11,7 @@ import { PlatService } from '../plat.service';
 })
 export class ListPlatComponent implements OnInit {
 listPlat : Plat[]=[]
-  constructor(private platservice :PlatService) { }
+  constructor(private platservice :PlatService,private roleguardservice : RoleGuardService) { }
 
   ngOnInit(): void {
     this.platservice.getAll().subscribe(
@@ -18,6 +19,11 @@ listPlat : Plat[]=[]
         this.listPlat=data;
       }
      )
+
+      if(localStorage.getItem("role")!=('Manager'|| 'Serveur')){
+      document.getElementById('idm').style.visibility="hidden";
+      document.getElementById('ids').style.visibility="hidden";
+      }  
   }
   getbyid(idplat:number){
     this.platservice.getbyid(idplat).subscribe(
@@ -27,17 +33,54 @@ listPlat : Plat[]=[]
     )
   }
   updatePlat(id:number,plat:Plat){
+    if(localStorage.getItem("role")==('Manager'|| 'Serveur')){
+
+      
     this.platservice.update(id,plat).subscribe(
       data=>(
         console.log(data)
       )
     )
-  }
+  }else{
+    Swal.fire({
+      icon: 'error',
+          text:'Accès refusé !',
+           } )
+  }}  
   deletePlat(id:number){
-    this.platservice.delete(id).subscribe(
-      data=>(
-        console.log(data)
-      )
-    )
-  }
+    if(localStorage.getItem("role")==('Manager'|| 'Serveur')){
+      Swal.fire({
+        title: 'Etes vous sur ?',
+        text: "Vous ne pourrez pas revenir en arrière!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Oui, Supprimer!'
+      }).then((result)=>{
+        if(result.value){
+          this.platservice.delete(id).subscribe(
+            data=>{
+      
+              if(data){
+                Swal.fire(
+                  'Plat supprimé ! ',
+                ).then(()=> window.location.reload())
+               
+              }
+            }
+          )
+        }
+      })
+
+  }else{
+    Swal.fire({
+      icon: 'error',
+          text:'Accès refusé !',
+          
+    })
+
+  }}
+
+
 }
